+++ git_commands.py
@@ -38,7 +38,7 @@ mods_load_order = [
     '.stash',
     '.statusbar',
     '.flow',
-    '.mv',
+    '.file',
 ]
 
 reload_mods = [mod for mod in sys.modules if mod[0:3] in ('git', 'Git') and sys.modules[mod] is not None]
@@ -65,7 +65,7 @@ try:
     from .git.diff import *  # noqa
     from .git.flow import *  # noqa
     from .git.history import *  # noqa
-    from .git.mv import *  # noqa
+    from .git.file import *  # noqa
     from .git.ignore import *  # noqa
     from .git.repo import *  # noqa
     from .git.stash import *  # noqa
@@ -82,7 +82,7 @@ except (ImportError, ValueError):
     from git.diff import *  # noqa
     from git.flow import *  # noqa
     from git.history import *  # noqa
-    from git.mv import *  # noqa
+    from git.file import *  # noqa
     from git.ignore import *  # noqa
     from git.repo import *  # noqa
     from git.stash import *  # noqa
